Updated type of RigLog id column to int8
git-svn-id: https://develop.eng.uts.edu.au/projects/personal/sahara/svn/SchedServer/trunk@2551 7ee224fa-57ca-4419-87f9-1f4314938d77

+++ InstallerFiles/migrationScriptV2ToV3Postgres.sql
@@ -32,7 +32,7 @@
         primary key (id)
     );
     create table rig_log (
-        id  bigserial not null unique,
+        id int8 not null unique,
         new_state varchar(20) not null,
         old_state varchar(20) not null,
         reason varchar(255) not null,
